MAkefile générique, fin de l'exo2

+++ executables/exo2.c
@@ -1,25 +1,26 @@
 //Exercice destiné à trouver les @ mémoires des différentes sections sur Linux
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 
-int test(int a, int b);
-int addition (int d, int e);
+int8_t test(int8_t a, int8_t b);
+int8_t addition (int8_t d, int8_t e);
 
 //Definition de variables dans chacune des 3 sections dont la variable est le nom
-const int rodata = 10;
-int data = 11;
-int bss;
+const int8_t rodata = 10;
+int8_t data = 11;
+int8_t bss;
 
 //Definition de fonctions imbriquées pour determiner le sens de la pile
-int addition(int d, int e) {
+int8_t addition(int8_t d, int8_t e) {
 	printf("addresse de d: %p, addresse de e: %p \n", &d, &e);
-	int f = d + e;
+	int8_t f = d + e;
 	return f;
 }
 
-int test(int a, int b) {
+int8_t test(int8_t a, int8_t b) {
 	printf("addrese de a: %p, addresse de b: %p \n", &a, &b);
-	int c = addition(3,2);
+	int8_t c = addition(3,2);
 	return c;
 }
 
@@ -34,3 +35,5 @@ int main() {
 	free(pointeur);
 	return 0;
 }
+//On obtient directement l'ordre des différentes sections  en récupérant et classant les adresses de rodata, data et bss, l'adresse de pointeur (malloc) pour le tas, les arguments des fonctions pour la pile (possibilité de les avoir en debuggant avec gdb.
+//Concernant le sens de la pile, ii suffit de regarder comment les adresses sont gérées dans les deux fonctions imbriquées addition et test comme vu dans le cours
